refactor(check): descend the thinking judge and static resolver to the substrate
nika-check sat at 14932 prod lines after hosting the judgment — the
thinking module's honest docs and teaching strings could never fit the
68-line headroom, and the wall exists precisely so a budget is not met
by deleting the prose it protects. The ADR-115 direction applies: the
module the wall-crate merely hosts descends to the analysis substrate.

thinking.rs and the shared static-value resolver (static_literal_of +
bare_static_ref, pure AST reads every lane reads) move to
nika-check-analyzer, which gains the nika-catalog edge for the seat
law's reasoning and known-model facts (L0 leaf, the same catalog the
hints lane reads). nika-check re-exports both at their historical root
paths — no lane's call site moved — and the CLI's thin wrapper, the
three fold sites, and the wiring pin are untouched. The judgment keeps
its via-default arm because the resolver moved with it. The public-api
baselines regenerate as purely additive at both crates (nika-check
grows ThinkingFinding + thinking_findings re-exports; static_literal_of
stays at the same public path as a re-export). Layering gate green;
nika-cli and nika-check both back under 15000.
