+++ crates/nika-check/src/lib.rs
@@ -151,7 +151,6 @@ mod schema_lint;
 mod schema_typing;
 mod secrets;
 pub mod slots;
-mod thinking;
 mod tools;
 pub mod trifecta;
 mod walk;
@@ -191,14 +190,15 @@ pub use schema_lint::SchemaLintFinding;
 pub use schema_typing::{SchemaTypeFinding, UnverifiableOutputRef};
 pub use secrets::{SecretEgress, SecretLeak};
 pub use slots::{MARKER_OPEN, SlotFinding};
-pub use thinking::{ThinkingFinding, thinking_findings};
 pub use tools::{MissingArg, UnknownArg, UnknownTool};
-pub use walk::{static_literal_of, static_read_paths};
+pub use walk::static_read_paths;
 
 // The analyzer's surface at the crate root — the same shape `nika-schema`
-// re-exported before the split (`analyze` · `AnalyzedWorkflow` · the
-// type-contract projections).
-pub use analyzer::{AnalyzedWorkflow, analyze, lowered_returns, returns_type};
+// re-exported before the split. `static_literal_of` and the thinking-seat
+// law descended to the substrate 2026-08-25 (the 15k wall · ADR-115's
+// direction); every lane's call path is unchanged.
+pub use analyzer::{AnalyzedWorkflow, analyze, lowered_returns, returns_type, static_literal_of};
+pub use analyzer::{ThinkingFinding, thinking_findings};
 
 /// The JSON contract version of [`CheckReport`] — bumped on any
 /// breaking field rename/removal so agent loops fail LOUDLY instead of
